Move coroutine test source to separate file
